src/algorithms/binarySearch/IntersectionOfTwoArrays.java
package algorithms.binarySearch;

import java.util.Arrays;
import java.util.HashSet;
import java.util.Set;

/**
 * size m > n, sort the bigger m array, use the smaller n array to do a binary search
 * and add existed results to result set.
 * time: O(mlogm + nlogm + 1) = O((m + n)logm)
 * space: O(n)
 */
public class IntersectionOfTwoArrays {
    public int[] intersection(int[] nums1, int[] nums2) {
        if (nums1.length > nums2.length) {
            intersection(nums2, nums1);
        }
        Set<Integer> set = new HashSet<>();
        Arrays.sort(nums2);
        for (Integer num : nums1) {
            if (binarySearch(nums2, num)) {
                set.add(num);
            }
        }
        int i = 0;
        int[] result = new int[set.size()];
        for (Integer num : set) {
            result[i++] = num;
        }
        return result;
    }
    private boolean binarySearch(int[] nums, int target) {
        int left = 0;
        int right = nums.length - 1;
        while (left <= right) {
            int mid = left + (right - left) / 2;
            if (nums[mid] == target) {
                return true;
            } else if (nums[mid] > target) {
                right = mid - 1;
            } else {
                left = mid + 1;
            }
        }
        return false;
    }
}